Move portable RegisteredWaitHandle implementation to shared ThreadPool.cs
